+++ test/index.js
@@ -1,19 +1,41 @@
 var assertDir = require('assert-dir-equal')
-var jstransformerTemplate = require('../')
 var Metalsmith = require('metalsmith')
+var rimraf = require('rimraf')
 
-function test (name, options) {
+/**
+ * Define a test case.
+ *
+ * @param name - The folder name for the test fixtures.
+ * @param plugins - An associative array of plugin key name, and options for it.
+ */
+function test (name, plugins) {
   /* globals it describe */
   it(name, function (done) {
-    Metalsmith('test/fixtures/' + name)
-      .use(jstransformerTemplate(options || {}))
-      .build(function (err) {
+    // Ensure we load the Metalsmith JSTransformer Layouts plugin.
+    plugins = plugins || {}
+    if (!plugins['..']) {
+      plugins['..'] = {}
+    }
+
+    // Construct Metalsmith with a clean build directory.
+    var testPath = 'test/fixtures/' + name
+    rimraf(testPath + '/build', function () {
+      var metalsmith = Metalsmith('test/fixtures/' + name)
+
+      // Load each Plugin.
+      for (var plugin in plugins || {}) {
+        metalsmith.use(require(plugin)(plugins[plugin]))
+      }
+
+      // Build the test.
+      metalsmith.build(function (err) {
         if (err) {
           return done(err)
         }
-        assertDir('test/fixtures/' + name + '/build', 'test/fixtures/' + name + '/expected')
+        assertDir(testPath + '/build', testPath + '/expected')
         return done()
       })
+    })
   })
 }
 
@@ -23,4 +45,7 @@ describe('metalsmith-jstransformer', function () {
   test('multiple')
   test('inherited')
   test('recursive')
+  test('jstransformer', {
+    'metalsmith-jstransformer': {}
+  })
 })