+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 - Renamed to `layouts`
 - Added inherited layout support
+- Added support for layouts with multiple template engines
 
 ## v0.0.1: 2015-07-03