+++ HISTORY.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## v0.1.0: Unreleased
+## v0.2.0: Unreleased
+
+- Added [`metalsmith-jstransformer`](https://github.com/RobLoach/metalsmith-jstransformer) test
+
+## v0.1.0: 2015-07-07
 
 - Renamed to `layouts`
 - Added inherited layout support